# openapi.json
{"openapi":"3.1.0","info":{"title":"FastAPI","version":"0.1.0"},"paths":{"/trips/":{"post":{"summary":"Create","operationId":"create_trips__post","parameters":[{"name":"start_time","in":"query","required":true,"schema":{"type":"string","title":"Start Time"}},{"name":"end_time","in":"query","required":true,"schema":{"type":"string","title":"End Time"}},{"name":"is_demand","in":"query","required":true,"schema":{"type":"boolean","title":"Is Demand"}},{"name":"start_floor","in":"query","required":true,"schema":{"type":"integer","maximum":7,"minimum":1,"title":"Start Floor"}},{"name":"end_floor","in":"query","required":true,"schema":{"type":"integer","maximum":7,"minimum":1,"title":"End Floor"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"get":{"summary":"Read","operationId":"read_trips__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/trips/{trip_id}/":{"put":{"summary":"Update","operationId":"update_trips__trip_id___put","parameters":[{"name":"trip_id","in":"path","required":true,"schema":{"type":"integer","title":"Trip Id"}},{"name":"end_floor","in":"query","required":true,"schema":{"type":"integer","maximum":7,"minimum":1,"title":"End Floor"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"summary":"Delete","operationId":"delete_trips__trip_id___delete","parameters":[{"name":"trip_id","in":"path","required":true,"schema":{"type":"integer","title":"Trip Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}